기타 수정

+++ src/main/java/com/sidam_backend/repo/NoticeRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 public interface NoticeRepository extends CrudRepository<Notice, Long> {
 
     @Query(value="SELECT * FROM notice_tbl WHERE store_id = :storeId AND id < :last AND valid = true ORDER BY id DESC LIMIT :cnt", nativeQuery = true)
-    List<Notice> selectAllAfterLast(int last, Long storeId, int cnt);
+    List<Notice> selectAllAfterLast(long last, Long storeId, int cnt);
 
     @Query(value="SELECT id FROM notice_tbl WHERE store_id = :storeId AND valid = true ORDER BY id DESC LIMIT 1", nativeQuery = true)
     Optional<Long> selectLastId(Long storeId);
